MCP_DeepResearch/requirements.txt
mcp
python-dotenv
google-generativeai
tavily-python
requests
markdown-it-py